chore: move reusable code to util files
