+++ pkg/groupsync/manytomanysyncer.go
@@ -234,11 +234,3 @@ func (f *ManyToManySyncer) targetUsers(ctx context.Context, sourceUsers []*User)
 	}
 	return targetUsers, merr
 }
-
-func userIDs(users []*User) []string {
-	ids := make([]string, 0, len(users))
-	for _, user := range users {
-		ids = append(ids, user.ID)
-	}
-	return ids
-}
