+++ pkg/groupsync/manytomanysyncer_test.go
@@ -15,10 +15,7 @@
 package groupsync
 
 import (
-	"context"
 	"fmt"
-	"sort"
-	"sync"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -1340,159 +1337,4 @@ func TestSyncAll(t *testing.T) {
 			}
 		})
 	}
-}
-
-type testReadWriteGroupClient struct {
-	groups          map[string]*Group
-	groupMembers    map[string][]Member
-	users           map[string]*User
-	descendantsErrs map[string]error
-	getGroupErrs    map[string]error
-	getMembersErrs  map[string]error
-	getUserErrs     map[string]error
-	setMembersErrs  map[string]error
-	mutex           sync.RWMutex
-}
-
-func (tc *testReadWriteGroupClient) Descendants(ctx context.Context, groupID string) ([]*User, error) {
-	tc.mutex.RLock()
-	defer tc.mutex.RUnlock()
-	if err, ok := tc.descendantsErrs[groupID]; ok {
-		return nil, err
-	}
-	return Descendants(ctx, groupID, tc.GetMembers)
-}
-
-func (tc *testReadWriteGroupClient) GetGroup(ctx context.Context, groupID string) (*Group, error) {
-	tc.mutex.RLock()
-	defer tc.mutex.RUnlock()
-	if err, ok := tc.getGroupErrs[groupID]; ok {
-		return nil, err
-	}
-	group, ok := tc.groups[groupID]
-	if !ok {
-		return nil, fmt.Errorf("group %s not found", groupID)
-	}
-	return group, nil
-}
-
-func (tc *testReadWriteGroupClient) GetMembers(ctx context.Context, groupID string) ([]Member, error) {
-	tc.mutex.RLock()
-	defer tc.mutex.RUnlock()
-	if err, ok := tc.getMembersErrs[groupID]; ok {
-		return nil, err
-	}
-	members, ok := tc.groupMembers[groupID]
-	if !ok {
-		return nil, fmt.Errorf("group %s not found", groupID)
-	}
-	return members, nil
-}
-
-func (tc *testReadWriteGroupClient) GetUser(ctx context.Context, userID string) (*User, error) {
-	tc.mutex.RLock()
-	defer tc.mutex.RUnlock()
-	if err, ok := tc.getUserErrs[userID]; ok {
-		return nil, err
-	}
-	user, ok := tc.users[userID]
-	if !ok {
-		return nil, fmt.Errorf("user %s not found", userID)
-	}
-	return user, nil
-}
-
-func (tc *testReadWriteGroupClient) SetMembers(ctx context.Context, groupID string, members []Member) error {
-	tc.mutex.Lock()
-	defer tc.mutex.Unlock()
-	if err, ok := tc.setMembersErrs[groupID]; ok {
-		return err
-	}
-	_, ok := tc.groupMembers[groupID]
-	if !ok {
-		return fmt.Errorf("group %s not found", groupID)
-	}
-	// sort members so we have deterministic ordering for comparisons
-	sort.Slice(members, func(i, j int) bool {
-		u1, err := members[i].User()
-		if err != nil {
-			// test is broken if this happens
-			panic(err)
-		}
-		u2, err := members[j].User()
-		if err != nil {
-			// test is broken if this happens
-			panic(err)
-		}
-		return u1.ID < u2.ID
-	})
-	tc.groupMembers[groupID] = members
-	return nil
-}
-
-type testGroupMapper struct {
-	m                   map[string][]string
-	allGroupIDsErr      error
-	containsGroupIDErrs map[string]error
-	mappedGroupIdsErr   map[string]error
-}
-
-func (tgm *testGroupMapper) AllGroupIDs(ctx context.Context) ([]string, error) {
-	if tgm.allGroupIDsErr != nil {
-		return nil, tgm.allGroupIDsErr
-	}
-	ids := make([]string, 0, len(tgm.m))
-	for id := range tgm.m {
-		ids = append(ids, id)
-	}
-	return ids, nil
-}
-
-func (tgm *testGroupMapper) ContainsGroupID(ctx context.Context, groupID string) (bool, error) {
-	if err, ok := tgm.containsGroupIDErrs[groupID]; ok {
-		return false, err
-	}
-	_, ok := tgm.m[groupID]
-	return ok, nil
-}
-
-func (tgm *testGroupMapper) MappedGroupIDs(ctx context.Context, groupID string) ([]string, error) {
-	if err, ok := tgm.mappedGroupIdsErr[groupID]; ok {
-		return nil, err
-	}
-	ids, ok := tgm.m[groupID]
-	if !ok {
-		return nil, fmt.Errorf("group %s not mapped", groupID)
-	}
-	return ids, nil
-}
-
-func (tgm *testGroupMapper) Mappings(ctx context.Context, groupID string) ([]Mapping, error) {
-	mappedGroupIDs, err := tgm.MappedGroupIDs(ctx, groupID)
-	if err != nil {
-		return nil, err
-	}
-	mappings := make([]Mapping, len(mappedGroupIDs))
-	for i, groupID := range mappedGroupIDs {
-		mappings[i] = Mapping{
-			GroupID: groupID,
-		}
-	}
-	return mappings, nil
-}
-
-type testUserMapper struct {
-	m                map[string]string
-	mappedUserIDErrs map[string]error
-}
-
-func (tum *testUserMapper) MappedUserID(ctx context.Context, userID string) (string, error) {
-	if err, ok := tum.mappedUserIDErrs[userID]; ok {
-		return "", err
-	}
-	id, ok := tum.m[userID]
-	if !ok {
-		return "", fmt.Errorf("user %s not mapped", userID)
-	}
-	return id, nil
 }
